requirements.txt
pandas==1.5.3
langchain
kor
openai
tiktoken
pinecone-client
pylint
pyyaml
streamlit